move logic of variant selection to main product details component
