+++ src/app/product/components/product-detail-page/product-detail-page.component.html
@@ -1,3 +1,3 @@
-<div *ngIf="(product$ | async)">
-  <app-product-details [product]="product$ | async"></app-product-details>
+<div *ngIf="product$ != null">
+  <app-product-details [product]="product$"></app-product-details>
 </div>
