+++ src/app/product/components/product-detail-page/product-detail-page.component.ts
@@ -12,29 +12,30 @@ import { ProductService } from './../../../core/services/product.service';
 })
 export class ProductDetailPageComponent implements OnInit {
   actionsSubscription: Subscription;
-  product$: Observable<Product>;
+  product$: Product = null;
   routeSubs: Subscription;
   productId: any;
 
   constructor(private productService: ProductService,
               private route: ActivatedRoute) {
-  };
 
   /**On Init
    * 1. Parse route params
    * 2. Retrive product id
    * 3. Ask for the product detail based on product id 
    * */
-  ngOnInit() {
     this.actionsSubscription = this.route.params.subscribe(
       (params: any) => {
         this.productId = params['id'];
-        this.product$ = this.productService.getProduct(this.productId);
+        this.productService
+          .getProduct(this.productId)
+          .subscribe(response => this.product$ = response);
      }
     );
+  };
 
-    this.product$.subscribe(res => console.log(res));
 
+  ngOnInit() {
   }
 
   /**
