+++ src/app/product/components/product-detail-page/product-details/product-details.component.html
@@ -2,11 +2,14 @@
   <app-image-container [product]="product" class="col-md-6"></app-image-container>
   <div class="col-md-6">
     <app-product-price-info [product]="product" class="row"></app-product-price-info>
-    <app-product-variants [product]="product" class="row"></app-product-variants>  
+    <app-product-variants
+      [customOptionTypesHash]="customOptionTypesHash"
+      [currentSelectedOptions]="currentSelectedOptions" 
+      (onOptionClickEvent)="onOptionClick($event)"
+      class="row"></app-product-variants>  
     <div>
       <button class="pdp-add-to-bag pdp-button">ADD TO BAG</button>
     </div>
     <app-product-description [product]="product" class="row"  ></app-product-description>  
   </div>
-
 </div>
